// src/common/types.ts
export interface AppMetadata {
  /* App name */
  name: string

  /* An SVG icon string or image url, with height set to 100%
    Note: `icon` is displayed on both mobile AND desktop. If `logo`
    below is provided then `icon` displays on mobile and `logo` on
    desktop.
  */
  icon?: string

  /* SVG logo (icon and text) string or image url, with width set to 100%
     Note: This will ONLY display on desktop. It is best used with wide
     format logos. Use `icon` for standard 40x40 icons.
  */
  logo?: string

  /* Description of app*/
  description?: string

  /* Url to a getting started guide for app */
  gettingStartedGuide?: string

  /* Url that points to more information about app */
  explore?: string

  /** When no injected wallets detected, recommend the user to install some*/
  recommendedInjectedWallets?: RecommendedInjectedWallets[]

  agreement?: TermsOfServiceAgreementOptions | null
}
// ...
export interface WalletModule {
  // The label of the wallet
  label: string
  /**
   * Gets the icon of the wallet
   * @returns
   */
  getIcon: () => Promise<string>
  /**
   * @returns the wallet interface associated with the module
   */
  getInterface: (helpers: GetInterfaceHelpers) => Promise<WalletInterface>
}

export type GetInterfaceHelpers = {
  chains: Chain[]
  appMetadata: AppMetadata | null
  BigNumber: typeof ethers.BigNumber
  EventEmitter: typeof EventEmitter
}

export type ChainId = string

export type DecimalChainId = number

export type RpcUrl = string

export type WalletInterface = {
  provider: EIP1193Provider
  instance?: unknown
}
// ...
export interface EIP1193Provider extends SimpleEventEmitter {
  on(event: 'connect', listener: ConnectListener): void
  on(event: 'disconnect', listener: DisconnectListener): void
  on(event: 'message', listener: MessageListener): void
  on(event: 'chainChanged', listener: ChainListener): void
  on(event: 'accountsChanged', listener: AccountsListener): void
  request(args: EthAccountsRequest): Promise<ProviderAccounts>
  request(args: EthBalanceRequest): Promise<Balance>
  request(args: EIP1102Request): Promise<ProviderAccounts>
  request(args: SelectAccountsRequest): Promise<ProviderAccounts>
  request(args: EIP3326Request): Promise<null>
  request(args: EIP3085Request): Promise<null>
  request(args: EthChainIdRequest): Promise<ChainId>
  request(args: EthSignTransactionRequest): Promise<string>
  request(args: EthSignMessageRequest): Promise<string>
  request(args: PersonalSignMessageRequest): Promise<string>
  request(args: EIP712Request): Promise<string>
  request(args: { method: string; params?: Array<unknown> }): Promise<unknown>
  disconnect?(): void
}

export enum ProviderRpcErrorCode {
  ACCOUNT_ACCESS_REJECTED = 4001,
  ACCOUNT_ACCESS_ALREADY_REQUESTED = -32002,
  UNAUTHORIZED = 4100,
  INVALID_PARAMS = -32602,
  UNSUPPORTED_METHOD = 4200,
  DISCONNECTED = 4900,
  CHAIN_DISCONNECTED = 4901,
  CHAIN_NOT_ADDED = 4902,
  DOES_NOT_EXIST = -32601,
  UNRECOGNIZED_CHAIN_ID = -32603
}

export interface Chain {
  /**
   * String indicating chain namespace.
   * Defaults to 'evm' but will allow other chain namespaces in the future
   */
  namespace?: 'evm'
  /* Hex encoded string, eg '0x1' for Ethereum Mainnet */
  id: ChainId
  /**
   * Recommended to include. Used for network requests
   * (eg Alchemy or Infura end point).
   * PLEASE NOTE: Some wallets require an rpcUrl, label,
   * and token for actions such as adding a new chain.
   * It is recommended to include rpcUrl, label,
   * and token for full functionality.
   */
  rpcUrl?: string
  /* Recommended to include. Used for display, eg Ethereum Mainnet */
  label?: string
  /* Recommended to include. The native token symbol, eg ETH, BNB, MATIC */
  token?: TokenSymbol
  /**
   * The color used to represent the chain and
   * will be used as a background for the icon
   */
  color?: string
  /* Svg string. The icon to represent the chain */
  icon?: string
  /* Related to ConnectionInfo from 'ethers/lib/utils' */
  providerConnectionInfo?: ConnectionInfo
  /* An optional public RPC used when adding a new chain config to the wallet */
  publicRpcUrl?: string
  /* Also used when adding a new config to the wallet */
  blockExplorerUrl?: string
}